refactor(chat): extract search/snippet/preview helpers to utils/chatSearch.js
Stage 2 of the Chat.jsx decomposition. Moves the chat search apparatus
(normalisation, match counting, message/conversation search-text builders,
system-message snippets, notification-event highlight resolution) plus the
conversation-list dedupe and last-message preview/hydration helpers into
utils/chatSearch.js, verbatim. Public helpers are imported back; internal
ones stay unexported.

No behaviour change: pure functions (plus the existing async preview
hydrator that runs inside the conversations queryFn) moved as-is.
Chat.jsx 3296 -> 2843 (3543 -> 2843 across stages 1+2).
ESLint clean (pre-existing dedupeMessages dep warning unchanged), build green.
